# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FurniCore ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PRODUCTS
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierBody"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /suppliers/{id}/quotes:
    get:
      operationId: getSupplierQuotes
      tags: [suppliers]
      summary: Get quotes for a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierQuote"

  # QUOTES (Price Lock Workflow)
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Quotes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierQuote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a quote (supplier submits price)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierQuote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierQuote"
  /quotes/{id}/lock:
    post:
      operationId: lockQuote
      tags: [quotes]
      summary: Lock quote price (supplier locks)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Locked quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierQuote"
  /quotes/{id}/approve:
    post:
      operationId: approveQuote
      tags: [quotes]
      summary: Admin approves locked quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierQuote"
  /quotes/{id}/pay:
    post:
      operationId: payQuote
      tags: [quotes]
      summary: Mark quote as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paid quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierQuote"

  # MANUFACTURING
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
      required:
        - name
        - email
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean

    # PRODUCTS
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        status:
          type: string
        rating:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - status
        - createdAt
        - updatedAt
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        contactPerson:
          type: string
      required:
        - name
        - email
    UpdateSupplierBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        contactPerson:
          type: string
        status:
          type: string
        rating:
          type: number

    # QUOTES
    SupplierQuote:
      type: object
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        inventoryItemId:
          type: ["integer", "null"]
        itemName:
          type: ["string", "null"]
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        lockedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - supplierId
        - supplierName
        - description
        - quantity
        - unitPrice
        - totalPrice
        - status
        - createdAt
        - updatedAt
    CreateQuoteBody:
      type: object
      properties:
        supplierId:
          type: integer
        inventoryItemId:
          type: integer
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        notes:
          type: string
        validUntil:
          type: string
      required:
        - supplierId
        - description
        - quantity
        - unitPrice

    # MANUFACTURING
...